batch/ingest/requirements.txt
google-cloud-storage==3.9.0
polars==1.38.1